PWA - UI (#84)
* UI - tools and action

* UI - pdf actions and tools

* UI - pdf info modal

* UI - style confirm and sign

* minor fix

* fix modal

+++ components/pdf/PdfPageActionTray.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="bg-paperdazgreen-300 sm:bg-transparent py-2 flex items-center text-black justify-between ml-[-2%] w-[104%]">
-    <div class="flex items-center gap-lg-4 flex-1 justify-between max-w-4xl px-lg-4 px-3">
-      <span class="font-bold">{{ isCreator ? file.fileName.length > 12 ? `${file.fileName.substr(0, 8)}...` : file.fileName : file.fileName }}</span>
+    <div class="flex items-center gap-lg-4 flex-1 justify-between px-lg-4 px-3 flex-wrap">
+      <span class="font-bold text-ellipsis whitespace-nowrap max-w-xs">{{ isCreator ? file.fileName.length > 12 ? `${file.fileName.substr(0, 8)}...` : file.fileName : file.fileName }}</span>
 
       <span class="hidden md:inline">
         <span class="circle circle-2 bg-[#757575]"></span>
@@ -125,9 +125,9 @@
       </span>
     </div>
 
-    <div class="hidden sm:flex items-center justify-end pe-lg-4">
+    <div class="hidden sm:flex items-center justify-end pe-lg-4" v-if="!isSign">
 
-      <button v-if="$auth.loggedIn" @click="saveChanges" :disabled="disablePublish"
+      <button v-if="$auth.loggedIn && !(isConfirm && !isCreator)" @click="saveChanges" :disabled="disablePublish"
         class="mr-2 text-xs text-white bg-paperdazgreen-400 rounded px-3 h-7 disabled:bg-gray-400 disabled:cursor-not-allowed">
         Done
       </button>
@@ -152,7 +152,7 @@
       use-default-button />
     <!-- v-model="showInitialModal" -->
 
-    <el-dialog :visible.sync="showPdfInfo" :append-to-body="true" :show-close="false" center width="100%" top="100vh" custom-class="-translate-y-full sm:hidden rounded"
+    <el-dialog :visible.sync="showPdfInfo" :append-to-body="true" :show-close="false" center width="100%" top="100vh" custom-class="-translate-y-full sm:hidden pdf-info-modal"
       class="bottom-0 overflow-hidden sm:hidden">
       <div class="w-full flex flex-col p-0 -mt-8 -mb-4">
         <div class="border-b flex items-center justify-between py-2">
@@ -296,9 +296,23 @@ export default Vue.extend({
     uploadedBy() {
       return this.file.uploadedBy || 'N/A'
     },
+    isSign() {
+      return String(this.file.fileAction).toLowerCase() === FileAction.SIGNED
+    },
+    userRole() {
+      return this.$auth.user.role;
+    },
+    isAgreedSign() {
+      return this.$store.state.agreeSign;
+    }
   },
   methods: {
     cancelPublish() {
+      // Toolbar function - cancelConfrim
+      if (this.isConfirm) {
+        window.location.assign('/dashboard')
+        return;
+      }
       if (this.upload_state) {
         this.$axios.delete(`/files/${this.file?.id}`)
           .then(() => {
